# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: ChiefVoice CRM API
  version: 1.0.0
  description: |
    AI-powered voice CRM backend.

    **Authentication:** All protected endpoints require a identity-provider-issued ID token.
    Pass it as `Authorization: Bearer <token>`.
    The SSE endpoint `/api/logs-stream` uses a short-lived ticket created by `/api/logs-stream/ticket`
    because browser EventSource cannot set custom Authorization headers. Other
    protected endpoints do not accept JWTs in query strings.

    **Roles:** `Organization Admin`, `Super Admin`, `Agent`, `Manager`.
    Platform-admin endpoints additionally require the email to be in `PLATFORM_ADMIN_EMAILS`.

  contact:
    name: ChiefVoice Support
    email: support@chiefvoice.ai
servers:
  - url: /api
    description: Current server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication ID token
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
...
    Lead:
      type: object
      properties:
        id:        { type: string }
        name:      { type: string }
        phone:     { type: string }
        email:     { type: string }
        status:    { type: string }
        score:     { type: number }
        assignedTo: { type: string }
        createdAt: { type: string, format: date-time }
...
    Document:
      type: object
      properties:
        id:       { type: string }
        name:     { type: string }
        type:     { type: string }
        uploadedAt: { type: string, format: date-time }
    DncEntry:
      type: object
      properties:
        id:     { type: string }
        phone:  { type: string }
        reason: { type: string }
...
security:
  - BearerAuth: []

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /leads:
    get:
      tags: [Leads]
      summary: List leads
      responses:
        '200':
          description: Lead list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Lead' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Lead' }
      responses:
        '201':
          description: Created lead
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lead' }
  /leads/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Leads]
      summary: Update lead
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Lead' }
      responses:
        '200': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Leads]
      summary: Delete lead
      responses:
        '200': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /search:
    post:
      tags: [Leads]
      summary: Search leads / records
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                q:      { type: string }
                table:  { type: string }
      responses:
        '200': { description: Search results }
...
  /enquiries:
    get:
      tags: [Enquiries]
      summary: List inbound enquiries
      responses:
        '200': { description: Enquiries }
    post:
      tags: [Enquiries]
      summary: Create enquiry
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '201': { description: Created }
  /enquiries/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Enquiries]
      summary: Update enquiry
      responses:
        '200': { description: Updated }
    delete:
      tags: [Enquiries]
      summary: Delete enquiry
      responses:
        '200': { description: Deleted }

  # ── Loans ─────────────────────────────────────────────────────────────────
  /loans:
    get:
      tags: [Loans]
      summary: List loan applications
      responses:
        '200': { description: Loans }
    post:
      tags: [Loans]
      summary: Create loan application
      requestBody:
        content:
          application/json:
            schema: { type: object }
      responses:
        '201': { description: Created }
  /loans/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [Loans]
      summary: Update loan
      responses:
        '200': { description: Updated }
    delete:
      tags: [Loans]
      summary: Delete loan
      responses:
        '200': { description: Deleted }

  # ── Catalog ───────────────────────────────────────────────────────────────
...
  /knowledge/documents:
    get:
      tags: [Knowledge]
      summary: List knowledge documents
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Document' } }
    post:
      tags: [Knowledge]
      summary: Add document (text body)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:    { type: string }
                content: { type: string }
      responses:
        '201': { description: Created }
  /knowledge/documents/upload:
    post:
      tags: [Knowledge]
      summary: Upload document file (PDF/DOCX)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201': { description: Uploaded }
  /knowledge/documents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Knowledge]
      summary: Delete document
      responses:
        '200': { description: Deleted }
...
  /compliance/dnc:
    get:
      tags: [Compliance]
      summary: List DNC (Do-Not-Call) entries
      responses:
        '200':
          description: DNC list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/DncEntry' } }
    post:
      tags: [Compliance]
      summary: Add number to DNC
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DncEntry' }
      responses:
        '201': { description: Added }
  /compliance/dnc/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Compliance]
      summary: Remove DNC entry
      responses:
        '200': { description: Removed }

  # ── Dashboard ─────────────────────────────────────────────────────────────
...